mkimage.sh: If urpmi is requested, configure media from within chroot
Sometimes, there's a bit of a mismatch between the host's urpmi
and the guest environment, such that it's unable to correctly add
the media and generate the urpmi.cfg. For example, it is not possible
to configure the rootfs urpmi.cfg for aarch64 from Mageia 6's urpmi.

In order to avoid that problem, we simply always execute urpmi.addmedia
from within the chroot. This gives us the side benefit of removing
the requirement for urpmi on the host when urpmi is going to be
installed into the rootfs.